backend/requirements.txt
django==5.2.10
django-cors-headers==4.9.0
openrouteservice==2.3.3
python-dotenv==1.2.1